requirements.txt
flask==3.0.3
mongoengine==0.29.1
flask-mongoengine-3==1.1.0
werkzeug==3.0.3
flask_cors
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
pyotp==2.9.0
qrcode==8.2
pillow==12.1.1
python-dotenv